+++ tests/test_plugins/yo.js
@@ -1,5 +1,0 @@
-exports.message = 'Yo!'
-exports.local = function($){
-	$.return()
-}
-module.parent.return()